% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID_NY_FL_tracking}
\alias{COVID_NY_FL_tracking}
\title{plotgardener example data for tracked COVID-19 cases in New York and Florida}
\format{
a dataframe with 3 columns
\describe{
\item{date}{The date of the case count.}
\item{state}{The state of the case count.
Either \code{"new york"} or \code{"florida"}.}
\item{caseIncrease}{The increase number of positive COVID-19 cases.}
}
}
\source{
Data was downloaded from The COVID Tracking Project
\url{https://covidtracking.com/}.
}
\usage{
data("COVID_NY_FL_tracking")
}
\description{
A timeline dataset tracking positive COVID-19 cases in
New York and Florida from 2020-01-29 to 2021-03-07.
}
\keyword{datasets}
