% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GM12878_HiC_10kb}
\alias{GM12878_HiC_10kb}
\title{BentoBox example GM12878 Hi-C data at 10 Kb resolution}
\format{
a 3-column data frame in sparse upper triangular format.
}
\usage{
data("GM12878_HiC_10kb")
}
\description{
A dataset containing interaction frequency matrix
counts along genomic coordinates in the
region chr21:28000000-30300000 according to the hg19
genome build. This data is from the GM12878
cell line.
}
\references{
Rao SS, Huntley MH, Durand NC, Stamenova EK,
Bochkov ID, Robinson JT, Sanborn AL, Machol I, Omer AD, Lander ES,
Aiden EL. A 3D map of the human genome at kilobase resolution reveals
principles of chromatin looping. Cell.
2014 Dec 18;159(7):1665-80. doi: 10.1016/j.cell.2014.11.021.
Epub 2014 Dec 11. Erratum in: Cell. 2015 Jul 30;162(3):687-8.
PMID: 25497547; PMCID: PMC5635824.
(\href{https://pubmed.ncbi.nlm.nih.gov/25497547/}{PubMed})
}
\keyword{datasets}
